refactor: moved the component icon-text-stat from the ui folder to the library folder.
