+++ src/library/component/atoms/icon-text-stat/icon-text-stat.module.scss
