+++ src/library/component/atoms/icon-text-stat/types/IProps.ts
